Migrate FtListLazyWrapper to the composition API
