+++ src/renderer/components/FtListLazyWrapper/FtListLazyWrapper.css
